Implementing TMX Polyline as Enemy Path

+++ quarkus-server/src/main/java/org/wanja/fatcat/map/LayerObjectPoint.java
@@ -1,6 +1,6 @@
 package org.wanja.fatcat.map;
 
 public class LayerObjectPoint {
-    public Integer x;
-    public Integer y;
+    public Double x;
+    public Double y;
 }
